corrected tests to have input complex tensors

+++ tf_complex/tests/activations_test.py
@@ -8,13 +8,13 @@ from tf_complex.activations import ModReLU, ComplexActivation
 class TestModReLU(TestCase):
     def test_init_and_call(self):
         modrelu = ModReLU()
-        inputs = tf.zeros([16, 32, 32, 8])
+        inputs = tf.zeros([16, 32, 32, 8], dtype=tf.complex64)
         outputs = modrelu(inputs)
         self.assertEqual(inputs.shape, outputs.shape)
 
 class TestComplexActivation(TestCase):
     def test_init_and_call(self):
-        inputs = tf.zeros([16, 32, 32, 8])
+        inputs = tf.zeros([16, 32, 32, 8], dtype=tf.complex64)
         for activation_str in ComplexActivation.activation_key_to_fun.keys():
             activation = ComplexActivation(activation_str)
             outputs = activation(inputs)
